Moved quicktype and ajv implementations to src/contrib. Updated and fixed tests.
