+++ src/shared/c8ypact/matcher.ts
@@ -66,6 +66,7 @@ export class C8yDefaultPactMatcher implements C8yPactMatcher {
 
     const parents = options?.parents ?? [];
     const strictMatching = options?.strictMatching ?? true;
+    debugger;
     const schemaMatcher =
       options?.schemaMatcher || C8yDefaultPactMatcher.schemaMatcher;
