+++ test/tsconfig.json
@@ -10,6 +10,7 @@
     "module": "Node16",
     "skipLibCheck": true,
     "composite": true,
+    "resolveJsonModule": true,
     "paths": {
       "cumulocity-cypress/*": ["./../src/*"],
       "cumulocity-cypress": ["./../src"]
